Added URL fecthing capability
When there is a URL in the message, soap is used to parse text out of it and feed that to the model. This can be used to summarize URL's, as an example.
Implemented "forget" feature. When the bot is referred and the first word is forget, message history is wiped